+++ lib/countly_flutter.dart
@@ -1,6 +1,7 @@
 export 'src/content_builder.dart';
 export 'src/countly_config.dart';
 export 'src/countly_flutter.dart';
+export 'src/feedback.dart';
 export 'src/remote_config.dart';
 export 'src/sessions.dart';
 export 'src/user_profile.dart';