+++ lib/src/countly_flutter.dart
@@ -7,11 +7,14 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:pedantic/pedantic.dart';
 
+import 'content_builder.dart';
 import 'content_builder_internal.dart';
 import 'countly_config.dart';
 import 'countly_state.dart';
 import 'device_id.dart';
 import 'device_id_internal.dart';
+import 'feedback.dart';
+import 'feedback_internal.dart';
 import 'remote_config.dart';
 import 'remote_config_internal.dart';
 import 'sessions.dart';
@@ -64,6 +67,7 @@ class Countly {
     _viewsInternal = ViewsInternal(_countlyState);
     _sessionsInternal = SessionsInternal(_countlyState);
     _contentBuilderInternal = ContentBuilderInternal(_countlyState);
+    _feedbackInternal = FeedbackInternal(_countlyState);
   }
   static final instance = _instance;
   static final _instance = Countly._();
@@ -87,6 +91,9 @@ class Countly {
 
   late final ContentBuilderInternal _contentBuilderInternal;
   ContentBuilderInternal get content => _contentBuilderInternal;
+
+  late final FeedbackInternal _feedbackInternal;
+  Feedback get feedback => _feedbackInternal;
 
   /// ignore: constant_identifier_names
   static const bool BUILDING_WITH_PUSH_DISABLED = false;
@@ -210,6 +217,18 @@ class Countly {
           Countly.instance._remoteConfigInternal.notifyVariantCallbacks(requestResult, error, id);
         } catch (e) {
           log('[FMethodCallH] $e', logLevel: LogLevel.ERROR);
+        }
+        break;
+      case 'feedbackCallback_onClosed':
+        if (_instance._feedbackInternal.feedbackCallback != null) {
+          _instance._feedbackInternal.feedbackCallback!.onClosed();
+          _instance._feedbackInternal.feedbackCallback = null;
+        }
+        break;
+      case 'feedbackCallback_onFinished':
+        if (_instance._feedbackInternal.feedbackCallback != null) {
+          _instance._feedbackInternal.feedbackCallback!.onFinished(call.arguments);
+          _instance._feedbackInternal.feedbackCallback = null;
         }
         break;
       case 'contentCallback':